+++ src/titiler/application/titiler/application/__init__.py
@@ -1,3 +1,3 @@
 """titiler.application"""
 
-__version__ = "0.5.1"
+__version__ = "0.6.0a1"